Add a table of group buffs that ZOMGBuffs doesn't normally track, but that conflict with ones that it does. Could also be done by adding it to the existing ids table for each individual buff, but I think doing it by category is cleaner.
Add Commanding Shout and Blood Pact for Stamina so that they show up instead of Power Word: Fortitude if present (BuffTehRaid is already avoiding buffing in that case, this just fixes the icons in the group list).

Move Dalaran Brilliance and Fel Intellegence to the new otherBuffs table, and remove special casing for them.

buffTypes is a superset of buffsLookup for all current uses of it, remove buffsLookup.

Don't remind DKs to buff Horn of Winter if they already have Battle Shout.