Some changes to get the thing working again with Rogues:
The ranged slot is gone--don't bother checking for it for temporary weapon enchants. Most special checks for rogue weapon buffs are there, except for one place where the syntax is in a state I don't care to touch just yet. 

Poisons are now spells and buffs. Switch all poison IDs over to the new ones, accounting for the new talented poisons. I've checked that this works on 2 rogues, one with and one without Leeching Poison. The colors that show in the menu now show 2 distinct colors for the damage vs. utility poisons. 

I've also added in the item check for Alchemist's Flask, as pointed out by Carsight.